+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed an issue where a model was incorrectly classified as an encoder model if it had
   no pipeline tag on the Hugging Face Hub and it relied on a custom implementation that
   isn't integrated into the `transformers` library.
-- Fixed an issue when a model config had no `pad_token_id`.
+- Fixed an issue when a model config had no `pad_token_id` and/or `eos_token_id`.
 
 ## [v16.11.0] - 2026-01-21